fix(audit-P1+P2): placeholder, eksik mesaj, TC algoritma, kaplama, format, tarih
Faz 2 — P1 hızlı 6 madde + P2-22.

P1-8: Placeholder "Sefa Yakut" → "Adınız"/"Soyadınız" (/odeme),
"Ad Soyad" (/adreslerim modal). Founder adının placeholder olarak
görünmesi müşteriye "veri sızdı mı?" hissi veriyordu.

P1-9: "Eksik: fatura bilgisi" → spesifik mesaj
- Bireysel + TC boş → "TC kimlik numarası"
- Bireysel + TC geçersiz → "TC kimlik (geçersiz numara — kontrol et)"
- Kurumsal eksik → "kurumsal fatura (VKN, vergi dairesi, ...)"
Kullanıcı tam olarak ne eksik söyleniyor.

P1-13: "Kaplama yok" → "Kaplamasız" (3 dosya)
- src/lib/pricing-engine/constants.ts ETIKET_COATINGS
- src/app/etiket/page.tsx COATINGS array
- src/components/pim/PimChat.tsx comparison literal
Pozitif form, daha doğal Türkçe.

P1-14: /sepet birim fiyat format /etiket ile aynı hassasiyetli
- decimal helper smart precision: exact 2-decimal ise "2,51", yoksa
  4-decimal "2,5079" göster
- TR + EN parity
- /etiket konfigüratör ile aynı formül (fmtUnit)

P1-15: /siparislerim tarih TR full month
- dateFmt month "short" → "long" ("17 May" → "17 Mayıs")
- TR'de ay kısaltma yaygın değil

P2-22: validation.ts TC error message
- "TC kimlik geçersiz (algoritma)" → "TC kimlik numarası hatalı — kontrol et"
- "algoritma" teknik dil, kullanıcı dostu değildi

Bonus: src/lib/format.ts — gelecek standardizasyon için locale-aware
helpers (formatCurrency, formatDate, formatDateWithWeekday).

Files:
- src/lib/format.ts (NEW)
- src/lib/validation.ts (algoritma msg)
- src/lib/pricing-engine/constants.ts (Kaplamasız)
- src/app/etiket/page.tsx (COATINGS)
- src/components/pim/PimChat.tsx (comparison)
- src/app/sepet/page.tsx (decimal smart precision)
- src/app/siparislerim/page.tsx (month long)
- src/app/odeme/page.tsx (placeholder + submitMissing detay)
- src/app/adreslerim/page.tsx (namePh)

+++ storefront/src/app/siparislerim/page.tsx
@@ -77,7 +77,9 @@ const COPY = {
     trackBtn: "Takip et",
     deliveredAt: "Teslim edildi",
     locale: "tr-TR",
-    dateFmt: { day: "numeric", month: "short", year: "numeric" } as const,
+    // Sefa 17 May P1-15: TR full month ("17 Mayıs 2026") — abbrev "17 May"
+    // yanlıştı (TR'de ay kısaltma yaygın değil)
+    dateFmt: { day: "numeric", month: "long", year: "numeric" } as const,
   },
   en: {
     eyebrow: "My account",
